styling change + git manifest cache + new git routes
